🚚 api: move route to src/certification/configuration context
