+++ api/src/certification/configuration/routes.js
@@ -1,0 +1,5 @@
+import * as complementaryCertification from './application/complementary-certification-route.js';
+
+const certificationConfigurationRoutes = [complementaryCertification];
+
+export { certificationConfigurationRoutes };
